New functions and handles. Performance improvments.Functions keys (top 5 keys you know) are now in implementation. You can only delete for now. Renaming soon I guess ^^

+++ src/core.h
@@ -15,12 +15,13 @@ extern uint16_t files_count;
 extern uint16_t current_file_index;
 extern uint16_t current_file_y;
 
+
 extern bool can_scroll_more;
 
 extern uint8_t files_type[1024];  
-extern uint16_t files_size[1024];  
+extern float files_size[1024];  
 extern char files_name[1024][9];  
 
-extern int16_t screen_scroll;
+extern uint16_t screen_scroll;
 
 #endif
